# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Reservas - Universidad",
        "description": "Backend para la gestión de reservas de recursos (salas, equipos, etc.) con autenticación JWT y roles admin/usuario.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/recursos": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Listar recursos activos",
                "description": "Obtiene la lista de recursos activos con opciones de filtrado por tipo y disponibilidad (cualquier usuario autenticado).",
                "operationId": "873571b93ee782b6112e2b8169b3c207",
                "parameters": [
                    {
                        "name": "tipo_recurso_id",
                        "in": "query",
                        "description": "Filtrar por ID del tipo de recurso",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "disponible",
                        "in": "query",
                        "description": "Filtrar por disponibilidad general (0=no disponible, 1=disponible)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de recursos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "tipo_recurso_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Aula 101"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "Aula con capacidad para 30 personas"
                                            },
                                            "ubicacion": {
                                                "type": "string",
                                                "example": "Edificio A, Primer Piso"
                                            },
                                            "capacidad": {
                                                "type": "integer",
                                                "example": 30
                                            },
                                            "disponibilidad_general": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "estado": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "tipo_recurso": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Aula"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Crear un nuevo recurso",
                "description": "Crea un nuevo recurso en el sistema (solo administradores).",
                "operationId": "ed6eb6ae6fa8ce1ae83d2e1dec8e5294",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo_recurso_id",
                                    "nombre"
                                ],
                                "properties": {
                                    "tipo_recurso_id": {
                                        "description": "ID del tipo de recurso (debe existir)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Aula Magna"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Auditorio principal con capacidad para 200 personas"
                                    },
                                    "ubicacion": {
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Edificio Principal, Tercer Piso"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 200
                                    },
                                    "disponibilidad_general": {
                                        "description": "Si no se especifica, por defecto es true",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recurso creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "tipo_recurso_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Aula Magna"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Auditorio principal con capacidad para 200 personas"
                                        },
                                        "ubicacion": {
                                            "type": "string",
                                            "example": "Edificio Principal, Tercer Piso"
                                        },
                                        "capacidad": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "disponibilidad_general": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "estado": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token JWT con el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}